% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataEmilie}
\alias{dataEmilie}
\title{dataEmilie}
\format{A data frame with 1,313 rows and 15 variables:
\itemize{
\item \code{tempssuividc10_1} A numeric vector.
\item \code{inddc10} A numeric vector.
\item \code{t1delai1} A numeric vector.
\item \code{t2delai1} A numeric vector.
\item \code{DEM1_10} A numeric vector.
\item \code{AGEENTRE} A numeric vector.
\item \code{sexe2} A numeric vector.
\item \code{CEP} A numeric vector.
\item \code{SPT1_1} A numeric vector.
\item \code{SPT1_2} A numeric vector.
\item \code{IADL4_1} A numeric vector.
\item \code{MMS_TW1} A numeric vector.
\item \code{ISA1_15} A numeric vector.
\item \code{BENTON1} A numeric vector.
\item \code{COD_W1} A numeric vector.
}}
\usage{
dataEmilie
}
\description{
Data.
}
\examples{
data("dataEmilie", package = "mla")
}
\keyword{datasets}
