% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriva.R
\name{deriva}
\alias{deriva}
\title{Numerical approch to derivate.}
\usage{
deriva(b, funcpa)
}
\arguments{
\item{b}{Value of parameters to be optimized over.}

\item{funcpa}{Function to be minimized (or maximized), with argument the vector of parameters over which minimization is to take place. It should return a scalar result.}
}
\value{
A list with the following elements:
\itemize{
\item \code{v}, the information score matrix;
\item \code{rl}, log-likelihood or likelihood of the model.
}
}
\description{
Function to return the first, second derivate and the information score matrix. The central finite-difference and forward finite-difference will be used.
}
\examples{
b <- 0.1
f <- function(b) {
  return((2 * b[1]**2 + 3 * b[1]))
}
d <- deriva(b = b, funcpa = f)
}
\references{
Donald W. Marquardt (1963). \emph{An algorithm for least-squares estimation of nonlinear parameters}. Journal of the Society for Industrial and Applied Mathematics, 11(2):431--441
}
\author{
Daniel Commenges
}
