% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mla-package.R
\docType{package}
\name{mla}
\alias{mla}
\alias{mla-package}
\title{An Algorithm for Least-squares Curve Fitting}
\description{
This algorithm provides a numerical solution to the
problem of minimizing a function. This is more efficient than
the Gauss-Newton-like algorithm when starting from points very
far from the final minimum. A new convergence test is
implemented (RDM) in addition to the usual stopping criterion:
stopping rule is when the gradients are small enough in the
parameters metric (GH-1G).
}
\references{
Donald W. Marquardt (1963). \emph{An algorithm for least-squares estimation of nonlinear parameters}. Journal of the Society for Industrial and Applied Mathematics, 11(2):431--441

Daniel Commenges,  H Jacqmin-Gadda, C. Proust, J. Guedj (2006). \emph{A Newton-like algorithm for likelihood maximization the robust-variance scoring algorithm}. arxiv:math/0610402v2
}
\author{
Alessandro Gasparini (ag475@leicester.ac.uk)
}
