% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.marqLevAlg.R
\name{summary.marqLevAlg}
\alias{summary.marqLevAlg}
\title{Summary of optimization.}
\usage{
\method{summary}{marqLevAlg}(object, digits = 8, ...)
}
\arguments{
\item{object}{An object of class \code{mla}.}

\item{digits}{Number of digits to print in outputs. Default value is 8.}

\item{...}{Unused.}
}
\description{
A short summary of parameters estimates via the \code{mla} algorithm.
}
\examples{
f1 <- function(b) {
  return(4 * (b[1] - 5)^2 + (b[2] - 6)^2)
}
test.marq <- marqLevAlg(
  b = c(8, 9), m = 2, maxiter = 100, epsa = 0.001, epsb = 0.001,
  epsd = 0.001, fn = f1
)

summary(test.marq)
}
\seealso{
mla

print.mla
}
\author{
Daniel Commenges

Melanie Prague

Amadou Diakite
}
\keyword{summary}
