% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weib}
\alias{weib}
\title{Simulated dataset for the Weibull function.}
\format{A data frame with 936 rows and 5 variables:
\itemize{
\item \code{c} A numeric vector.
\item \code{t0} Entry time.
\item \code{t1} Left border of interval censored.
\item \code{t2} Right border of interval censored.
\item \code{ve} A numeric vector.
}}
\usage{
weib
}
\description{
A simulated dataset for the Weibull function. Contains 936 rows and 5 columns.
}
\examples{
data("weib", package = "mla")
}
\keyword{datasets}
